Testing the load characters list and decode to the Character Model
